requirements.txt
# Основные зависимости
python-dotenv==1.0.0
pyTelegramBotAPI==4.14.1
aiohttp==3.9.1
pydantic-settings==2.6.1 
# Дополнительные утилиты
python-telegram-bot==20.7
requests==2.31.0
